+++ src/vmware/azext_vmware/_params.py
@@ -24,6 +24,7 @@ def load_arguments(self, _):
         c.argument('cluster_size', help='Number of hosts for the default management cluster. Minimum of 3 and maximum of 16.')
         c.argument('internet', help='Connectivity to internet. Specify "Enabled" or "Disabled".')
         c.argument('yes', help='Delete without confirmation.')
+        c.argument('identity', help='Enable a system assigned identity.', arg_type=get_enum_type(['SystemAssigned', 'None']))
 
     with self.argument_context('vmware cluster') as c:
         c.argument('name', options_list=['--name', '-n'], help='Name of the cluster.')
@@ -39,7 +40,7 @@ def load_arguments(self, _):
         c.argument('nsxt_password', help='NSX-T Manager password.')
         c.argument('accept_eula', help='Accept the end-user license agreement without prompting.')
         c.argument('network_block', help='A subnet at least of size /22. Make sure the CIDR format is conformed to (A.B.C.D/X) where A,B,C,D are between 0 and 255, and X is between 0 and 22.')
-        c.argument('identity', help='Enable a system assigned identity.')
+        c.argument('mi_system_assigned', help='Enable a system assigned identity.', arg_type=get_enum_type(['SystemAssigned', 'None']), deprecate_info=c.deprecate(redirect='--identity', hide=True))
 
     with self.argument_context('vmware private-cloud show') as c:
         c.argument('name', options_list=['--name', '-n'], help='Name of the private cloud.')