+++ src/vmware/azext_vmware/tests/latest/test_vmware_scenario.py
@@ -95,20 +95,17 @@ class VmwareScenarioTest(ScenarioTest):
         # delete the private cloud
         self.cmd('vmware private-cloud delete -g {rg} -n {privatecloud} --yes')
 
-        # set managed identity
-        self.cmd('vmware private-cloud identity assign -g {rg} -c {privatecloud}')
-
-        # remove managed identity
-        self.cmd('vmware private-cloud identity remove -g {rg} -c {privatecloud}')
-
-        # show managed identity
-        self.cmd('vmware private-cloud identity show -g {rg} -c {privatecloud}')
-
         # enable cmk encryption
         self.cmd('az vmware private-cloud enable-cmk-encryption -c {privatecloud} -g {rg} --enc-kv-key-name test-key-name --enc-kv-key-version 1 --enc-kv-url test-url')
 
         # disable cmk encyrption
         self.cmd('az vmware private-cloud disable-cmk-encryption -c {privatecloud} -g {rg} --yes')
+
+        # enable system assigned identity
+        self.cmd('az vmware private-cloud update -n {privatecloud} -g {rg} --identity SystemAssigned')
+
+        # disable system assigned identity
+        self.cmd('az vmware private-cloud update -n {privatecloud} -g {rg} --identity None')
 
         count = len(self.cmd('vmware private-cloud list -g {rg}').get_output_in_json())
         self.assertEqual(count, 1, 'private cloud count expected to be 1')