+++ src/vmware/HISTORY.md
@@ -1,13 +1,15 @@
 # Release History
 
-## 4.0.3 (2022-07)
+## 4.1.0 (2022-07)
 
+- Add new parameter `--identity` to `az vmware private-cloud create`
+- Add new parameter `--identity` to `az vmware private-cloud update`
 - Add `az vmware private-cloud enable-cmk-encryption`
 - Add `az vmware private-cloud disable-cmk-encryption`
 - Deprecate `az vmware private-cloud add-cmk-encryption`
 - Deprecate `az vmware private-cloud delete-cmk-encryption`
-- Removed `--mi-system-assigned` parameter from `az vmware private-cloud identity assign`
-- Renamed `--mi-system-assigned` parameter from `az vmware private-cloud create` to `--identity`
+- Deprecate `--mi-system-assigned` parameter from `az vmware private-cloud create`
+- Deprecate `az vmware private-cloud identity` command group
 
 ## 4.0.2 (2022-02)